refactor: improve file structure consistency and organization
File Content Components:
- Extract SkillsCodeContent from inline definition to separate file
- Rename AboutContent.tsx to AboutCodeContent.tsx for naming consistency
- All fileContent components now follow *CodeContent.tsx naming pattern

This improves:
- Code organization and maintainability
- Consistency across component naming
- Easier navigation and understanding of codebase structure

Documentation Updates:
- Update ARCHITECTURE.md with correct file names and structure
- Update DEVELOPMENT.md with improved examples for adding new sections
- Include proper import examples and component usage patterns
